+++ CHANGELOG.rst
@@ -1,4 +1,5 @@
 Current
 =======
 
+- Context: Remove experiment constructor. Remove :code:`push_context`.
 - Remove :code:`Experiment.set_update`. The more general :code:`Experiment.command` should be used.